refactor: move src/types/api/auth to shared/types/form/form_auth
