+++ src/public/services/UpdateFormService.ts
@@ -8,18 +8,18 @@ import {
 import {
   AdminFormDto,
   AdminFormViewDto,
+  EndPageUpdateDto,
   FormSettings,
+  LogicDto,
+  PermissionsUpdateDto,
   SettingsUpdateDto,
-} from '../../../shared/types/form/form'
-import { LogicDto } from '../../../shared/types/form/form_logic'
+  StartPageUpdateDto,
+} from '../../../shared/types/form'
 import { SubmissionResponseDto } from '../../../shared/types/submission'
 import {
   EmailSubmissionDto,
   EncryptSubmissionDto,
-  EndPageUpdateDto,
   FormUpdateParams,
-  PermissionsUpdateDto,
-  StartPageUpdateDto,
 } from '../../types/api'
 import { createEmailSubmissionFormData } from '../utils/submission'
