+++ src/types/api/form_auth.ts
@@ -1,0 +1,1 @@
+export * from '../../../shared/types/form/form_auth'
